MHR/PPR API staff reg tables return staff account ID.
